Refactor screen transition
