+++ webgame-game/src/ui/mod.rs
@@ -1,13 +1,16 @@
 pub mod menu_button;
+pub mod screen_transition;
 
 use bevy::prelude::*;
 use menu_button::MenuButtonPlugin;
+use screen_transition::ScreenTransitionPlugin;
 
 /// Manages UI logic and components.
 pub struct UIPlugin;
 
 impl Plugin for UIPlugin {
     fn build(&self, app: &mut App) {
-        app.add_plugins(MenuButtonPlugin);
+        app.add_plugins(MenuButtonPlugin)
+            .add_plugins(ScreenTransitionPlugin);
     }
 }
